Revert "Merge branch 'main' into HUYPS39177"
This reverts commit 287a1ed688d773ce9260ebfbb5ae054188649e59, reversing
changes made to 61007cc0f5ea6b6eebcd8b5c0be8d49ca11a8ecf.

+++ AOS_BE/src/main/java/com/aos/AOSBE/Repository/ReviewsRepository.java
@@ -26,7 +26,7 @@ public interface ReviewsRepository extends JpaRepository<Reviews, Integer>, JpaS
     Integer countReviewsByProductItemIdAndCreateAtBetween(Integer productItemId, LocalDateTime startDate, LocalDateTime endDate);
 
     @Query("SELECT AVG(r.rating) FROM Reviews r WHERE r.productItems.id = :productItemId")
-    Double findAverageRatingByProductItemId(@Param("productItemId") Integer productItemId);
+    Double findAverageRatingByProductItemId(@Param("productItemId") Long productItemId);
     @Query("SELECT COUNT(r) FROM Reviews r WHERE r.productItems.id = :productItemId")
     Long countByProductItemId(@Param("productItemId") Long productItemId);
     @Query("SELECT COUNT(r) > 0 FROM Reviews r WHERE r.accounts.id = :accountId AND r.productItems.id = :productItemId AND r.orders.id = :orderId")
